fix: address Copilot PR review comments
- prism_web_plugin.dart: track _wasmLoadingFuture to guard concurrent callers in ensureWasmLoaded; reset on error to allow retry
- main.dart: wrap _engine.shutdown() with unawaited() in dispose(); import dart:async
- FlutterMethodHandler.kt: cast rotationSpeed/fps to Double in getState — StandardMessageCodec on Android rejects Float
- PrismPlatformView.kt: call stopRendering() on render-loop exception to release GPU resources and detach bridge (was only setting running = false)
- PrismPlatformView.swift: move isInitialized = true into success path of configureDemo callback; reset to false and show error label on failure or nil handle; expose isReady computed property (demoHandle != nil)
- PrismFlutterPlugin.swift: isInitialized channel method now returns activePlatformView?.isReady ?? false instead of always true
- prism_render_view_web.dart: add _initializing guard; move _initialized = true after successful WASM load; wrap in try/catch/finally to log errors without silently blocking retries

+++ prism-flutter/src/commonMain/kotlin/com/hyeonslab/prism/flutter/FlutterMethodHandler.kt
@@ -35,9 +35,9 @@ class FlutterMethodHandler(private val bridge: PrismBridge) {
       "getState" -> {
         val state = bridge.state.value
         mapOf(
-          "rotationSpeed" to state.rotationSpeed,
+          "rotationSpeed" to state.rotationSpeed.toDouble(),
           "isPaused" to state.isPaused,
-          "fps" to state.fps,
+          "fps" to state.fps.toDouble(),
         )
       }
       "shutdown" -> {
